Update KinoWheels_JSON.ino
A long battle to figure out how to get the device to respond to a request and send data back to the device / network card that sent the request. Had to figure out how to get a UDP string parse it through JSON, and then figure out what the JSON String wanted to reset the data stream to 0,0,0 or send or stop sending data.